fixing ddl to remove decimals

+++ src/main/resources/ddls/tpcc-cockroachdb-ddl.sql
@@ -6,9 +6,9 @@ CREATE TABLE order_line (
   ol_number int NOT NULL,
   ol_i_id int NOT NULL,
   ol_delivery_d timestamp NULL DEFAULT NULL,
-  ol_amount decimal(6,2) NOT NULL,
+  ol_amount float NOT NULL,
   ol_supply_w_id int NOT NULL,
-  ol_quantity decimal(2,0) NOT NULL,
+  ol_quantity int NOT NULL,
   ol_dist_info char(24) NOT NULL,
   PRIMARY KEY (ol_w_id,ol_d_id,ol_o_id,ol_number)
 );
@@ -25,8 +25,8 @@ DROP TABLE IF EXISTS stock;
 CREATE TABLE stock (
   s_w_id int NOT NULL,
   s_i_id int NOT NULL,
-  s_quantity decimal(4,0) NOT NULL,
-  s_ytd decimal(8,2) NOT NULL,
+  s_quantity int NOT NULL,
+  s_ytd float NOT NULL,
   s_order_cnt int NOT NULL,
   s_remote_cnt int NOT NULL,
   s_data varchar(50) NOT NULL,
@@ -50,8 +50,8 @@ CREATE TABLE oorder (
   o_id int NOT NULL,
   o_c_id int NOT NULL,
   o_carrier_id int DEFAULT NULL,
-  o_ol_cnt decimal(2,0) NOT NULL,
-  o_all_local decimal(1,0) NOT NULL,
+  o_ol_cnt int NOT NULL,
+  o_all_local int NOT NULL,
   o_entry_d timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
   PRIMARY KEY (o_w_id,o_d_id,o_id),
   UNIQUE (o_w_id,o_d_id,o_c_id,o_id)
@@ -65,7 +65,7 @@ CREATE TABLE history (
   h_d_id int NOT NULL,
   h_w_id int NOT NULL,
   h_date timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
-  h_amount decimal(6,2) NOT NULL,
+  h_amount float NOT NULL,
   h_data varchar(24) NOT NULL
 );
 
@@ -74,12 +74,12 @@ CREATE TABLE customer (
   c_w_id int NOT NULL,
   c_d_id int NOT NULL,
   c_id int NOT NULL,
-  c_discount decimal(4,4) NOT NULL,
+  c_discount float NOT NULL,
   c_credit char(2) NOT NULL,
   c_last varchar(16) NOT NULL,
   c_first varchar(16) NOT NULL,
-  c_credit_lim decimal(12,2) NOT NULL,
-  c_balance decimal(12,2) NOT NULL,
+  c_credit_lim float NOT NULL,
+  c_balance float NOT NULL,
   c_ytd_payment float NOT NULL,
   c_payment_cnt int NOT NULL,
   c_delivery_cnt int NOT NULL,
@@ -99,8 +99,8 @@ DROP TABLE IF EXISTS district;
 CREATE TABLE district (
   d_w_id int NOT NULL,
   d_id int NOT NULL,
-  d_ytd decimal(12,2) NOT NULL,
-  d_tax decimal(4,4) NOT NULL,
+  d_ytd float NOT NULL,
+  d_tax float NOT NULL,
   d_next_o_id int NOT NULL,
   d_name varchar(10) NOT NULL,
   d_street_1 varchar(20) NOT NULL,
@@ -116,7 +116,7 @@ DROP TABLE IF EXISTS item;
 CREATE TABLE item (
   i_id int NOT NULL,
   i_name varchar(24) NOT NULL,
-  i_price decimal(5,2) NOT NULL,
+  i_price float NOT NULL,
   i_data varchar(50) NOT NULL,
   i_im_id int NOT NULL,
   PRIMARY KEY (i_id)
@@ -125,8 +125,8 @@ CREATE TABLE item (
 DROP TABLE IF EXISTS warehouse;
 CREATE TABLE warehouse (
   w_id int NOT NULL,
-  w_ytd decimal(12,2) NOT NULL,
-  w_tax decimal(4,4) NOT NULL,
+  w_ytd float NOT NULL,
+  w_tax float NOT NULL,
   w_name varchar(10) NOT NULL,
   w_street_1 varchar(20) NOT NULL,
   w_street_2 varchar(20) NOT NULL,
